Logic: Ubuntu: This commit provides an example .desktop file
This .desktop will launch the Saleae Logic application using the icons
provided in the previous commit. The application will need to be installed
to a common path such as `/usr/bin` with the name `logic`. Once this is
completed this .desktop file must be copied to `/usr/share/applications`

Enjoy!

Signed-off-by: Eric Schikschneit <eric.schikschneit@novatechautomation.com>